Split stitching
