+++ examples/src/main/scala/example/stitching/ExampleApp.scala
@@ -5,7 +5,7 @@ import caliban.quick._
 import caliban.schema.ArgBuilder.auto._
 import caliban.schema.Schema.auto._
 import caliban.schema._
-import caliban.tools.stitching.{ HttpRequest, RemoteResolver, RemoteSchemaResolver, ResolveRequest }
+import caliban.stitching.{ HttpRequest, RemoteResolver, RemoteSchemaResolver, ResolveRequest }
 import caliban.tools.{ Options, RemoteSchema, SchemaLoader }
 import sttp.client4.Backend
 import sttp.client4.httpclient.zio._
