+++ stitching/src/main/scala/caliban/stitching/package.scala
@@ -1,6 +1,4 @@
-package caliban.tools
-
-import caliban.{ CalibanError, ResponseValue }
+package caliban
 
 import sttp.client4._
